Add files via upload
Added three files that begin to make comparisons between:
- non-CHERI code compiled in a non-CHERI environment (fakedb_vanilla.c)
- non-CHERI code compiled in QEMU morello-purecap (fakedb_cherryenv.c)
- CHERI code compiled in QEMU morello-purecap (fakedb_morello.c)

This is a fake DB consisting of an employee and an order (and a money struct and a customer struct, belonging to the employee and the order respectively).

Absolutely nothing interesting is happening in these files - yet (semi-interesting stuff coming soon). 

The first two play around with primitive data types and observe their individual values when overflowed (only the int so far).

The third uses pp_cap to output the characteristics of the employee and order, which have been made pointers for the purpose of the exercise, but otherwise not changed.